refactor: paymentAmount 로직을 useOrderProuducts훅으로 분리
